6.1 update- Upside down triangles fix, Load image fix, drawing on image fix, stroke weight fix
Unsure if stroke weight is totally fixed

+++ Drawing_App/WeightCommand.pde
@@ -4,8 +4,8 @@
 
 public class WeightCommand implements Command{
   
-  private int newWeight = 1;
-  private int prevWeight;
+  private float newWeight = 1;
+  private float prevWeight;
  
   public WeightCommand(int w){
     this.prevWeight = newWeight;
